Fix bug in task table sort
